refactor: move ccsa-admin-tab to components/softball/admin-tabs/ccsa-sync
